+++ src/Models/User.php
@@ -9,6 +9,7 @@ use Hirtz\Skeleton\Behaviors\TimestampBehavior;
 use Hirtz\Skeleton\Behaviors\TrailBehavior;
 use Hirtz\Skeleton\Db\ActiveRecord;
 use Hirtz\Skeleton\Helpers\SecretKey;
+use Hirtz\Skeleton\Models\Breadcrumb;
 use Hirtz\Skeleton\Models\Interfaces\CustomAttributeInterface;
 use Hirtz\Skeleton\Models\Interfaces\SearchableInterface;
 use Hirtz\Skeleton\Models\Interfaces\StatusAttributeInterface;
@@ -512,6 +513,11 @@ class User extends ActiveRecord implements CustomAttributeInterface, IdentityInt
     public function getAdminRoute(): array
     {
         return $this->id ? ['/admin/user/update', 'id' => $this->id] : ['/admin/user/index'];
+    }
+
+    public function getAdminIndexBreadcrumb(): Breadcrumb
+    {
+        return new Breadcrumb(Yii::t('skeleton', 'COMMON_USERS'), ['/admin/user/index']);
     }
 
     public function getPermissionName(): string